requirements.txt
opencv-contrib-python>=4.5.0
numpy>=1.19.0
pillow>=8.0.0
pandas>=1.2.0
pyttsx3>=2.90